Split Ensemble and Cluster importance, fixed doctstring.
