+++ test/test_ensemble_importance.py
@@ -7,7 +7,7 @@ from sklearn.cluster import FeatureAgglomeration
 from sklearn.feature_extraction import image
 from sklearn.linear_model import LassoCV, MultiTaskLassoCV
 
-from hidimstat import CluVI, DesparsifiedLasso, EnCluVI
+from hidimstat import ClusterImportance, DesparsifiedLasso, EnsembleImportance
 from hidimstat._utils.scenario import (
     multivariate_simulation,
     multivariate_simulation_spatial,
@@ -44,102 +44,6 @@ def spatially_relaxed_fdp_power(
     fdp = len(false_positive) / len(selected_ids)
     power = len(true_positive) / len(beta_ids)
     return fdp, power
-
-
-def test_cluvi_spatial():
-    """
-    Test CluVI on a 2D spatial simulation. Testing for support recovery methods using
-    clustering is challenging as clusters that intersect the true support can also
-    include non-support features, rapidly increasing false positives. To address this,
-    we introduce a spatial relaxation in the evaluation metrics.
-
-     - Test that the spatially relaxed FDP is below a specified FDR threshold (0.1).
-     - Test that the statistical power is above a specified threshold (0.8).
-    """
-    n_samples = 400
-    shape = (10, 10)
-    n_features = shape[1] * shape[0]
-    roi_size = 2  # size of the edge of the four predictive regions
-    signal_noise_ratio = 32.0  # noise standard deviation
-    smooth_X = (
-        0.2  # level of spatial smoothing introduced by the Gaussian filter
-    )
-
-    fp_list = []
-    power_list = []
-    for seed in range(10):
-        # generating the data
-        X_init, y, beta, _ = multivariate_simulation_spatial(
-            n_samples, shape, roi_size, signal_noise_ratio, smooth_X, seed=seed
-        )
-
-        y = y - np.mean(y)
-        X_init = X_init - np.mean(X_init, axis=0)
-
-        n_clusters = 50
-        connectivity = image.grid_to_graph(n_x=n_features, n_y=1, n_z=1)
-        clustering = FeatureAgglomeration(
-            n_clusters=n_clusters, connectivity=connectivity, linkage="ward"
-        )
-        estimator = LassoCV(
-            max_iter=1000, tol=0.0001, eps=0.01, fit_intercept=False
-        )
-        cluvi = CluVI(
-            vi_estimator=DesparsifiedLasso(estimator=estimator),
-            clustering=clustering,
-            random_state=seed,
-        )
-        cluvi.fit_importance(X_init, y)
-        fwer = 0.1
-        selected = cluvi.fwer_selection(fwer=fwer, two_tailed_test=False)
-
-        fdp, power = spatially_relaxed_fdp_power(
-            selected=selected,
-            ground_truth=beta,
-            roi_size=roi_size,
-            spatial_tolerance=3,
-            shape=shape,
-        )
-        fp_list.append(int(fdp > 0))
-        power_list.append(power)
-    assert np.mean(power_list) >= 0.5
-    assert np.mean(fp_list) <= fwer
-
-
-def test_cluvi_independence():
-    """Test that CluVI works with repeated calls
-    non-regression test for #425
-    """
-    n_samples = 50
-    shape = (20, 20)
-    roi_size = 4  # size of the edge of the four predictive regions
-    X_init, y, _, _ = multivariate_simulation_spatial(
-        n_samples, shape, roi_size, signal_noise_ratio=10.0, smooth_X=1
-    )
-    alpha = 0.05  # alpha is the significance level for the statistical test
-    n_clusters = 50
-    connectivity = image.grid_to_graph(n_x=shape[0], n_y=shape[1])
-    ward = FeatureAgglomeration(
-        n_clusters=n_clusters, connectivity=connectivity, linkage="ward"
-    )
-    vi_estimator = DesparsifiedLasso(
-        estimator=LassoCV(
-            max_iter=1000, tol=0.0001, eps=0.01, fit_intercept=False
-        )
-    )
-
-    c1 = CluVI(
-        vi_estimator=vi_estimator,
-        clustering=ward,
-    )
-    c1.fit_importance(X_init, y)
-    s1 = c1.fwer_selection(alpha, n_tests=n_clusters)
-    c2 = CluVI(
-        vi_estimator=vi_estimator, clustering=ward, cluster_bootstrap_size=0.5
-    )
-    c2.fit_importance(X_init, y)
-    s2 = c2.fwer_selection(alpha, n_tests=n_clusters)
-    assert np.sum(s2) > np.sum(s1 / 2)
 
 
 def test_encluvi_spatial():
@@ -182,15 +86,18 @@ def test_encluvi_spatial():
         estimator = LassoCV(
             max_iter=1000, tol=0.0001, eps=0.01, fit_intercept=False
         )
-        cluvi = EnCluVI(
-            vi_estimator=DesparsifiedLasso(estimator=estimator),
-            clustering=clustering,
-            n_bootstraps=5,
+        encludl = EnsembleImportance(
+            vim=ClusterImportance(
+                vim=DesparsifiedLasso(estimator=estimator),
+                clustering=clustering,
+                random_state=seed,
+            ),
+            n_repeats=5,
             random_state=seed,
         )
-        cluvi.fit_importance(X_init, y)
+        encludl.fit_importance(X_init, y)
         fwer = 0.1
-        selected = cluvi.fwer_selection(fwer=fwer, two_tailed_test=False)
+        selected = encludl.fwer_selection(fwer=fwer, two_tailed_test=False)
 
         fdp, power = spatially_relaxed_fdp_power(
             selected=selected,
@@ -203,67 +110,6 @@ def test_encluvi_spatial():
         power_list.append(power)
     assert np.mean(power_list) >= 0.5
     assert np.mean(fp_list) <= fwer + tol
-
-
-def test_cluvi_temporal():
-    """
-    Testing the procedure on two simulations with a 1D data structure and
-    with n << p: with a temporal dimension. The support is connected and
-    of size 10, it must be recovered with a small spatial tolerance
-    parametrized by `margin_size`.
-    """
-    n_samples, n_features, n_target = 100, 500, 3
-    support_size = 10
-    signal_noise_ratio = 50.0
-    rho_serial = 0.9
-    rho_data = 0.9
-    n_clusters = 100
-    margin_size = 5
-    extended_support = support_size + margin_size
-    test_tol = 0.05
-
-    fdp_list = []
-    power_list = []
-    for seed in range(10):
-        X, y, _, _ = multivariate_simulation(
-            n_samples=n_samples,
-            n_features=n_features,
-            n_targets=n_target,
-            support_size=support_size,
-            signal_noise_ratio=signal_noise_ratio,
-            rho_serial=rho_serial,
-            rho=rho_data,
-            shuffle=False,
-            continuous_support=True,
-            seed=seed,
-        )
-
-        connectivity = image.grid_to_graph(n_x=n_features, n_y=1, n_z=1)
-        ward = FeatureAgglomeration(
-            n_clusters=n_clusters, connectivity=connectivity, linkage="ward"
-        )
-
-        cluvi = CluVI(
-            vi_estimator=DesparsifiedLasso(
-                estimator=MultiTaskLassoCV(max_iter=1000)
-            ),
-            clustering=ward,
-            random_state=seed,
-        )
-        cluvi.fit_importance(X, y)
-
-        alpha = 0.05
-        selected = cluvi.fdr_selection(fdr=alpha, two_tailed_test=False)
-        gt_mask = np.zeros(n_features, dtype=int)
-        gt_mask[:extended_support] = 1
-        fdp, power = fdp_power(
-            selected=selected,
-            ground_truth=gt_mask,
-        )
-        fdp_list.append(fdp)
-        power_list.append(power)
-    assert np.mean(power_list) >= 0.5 - test_tol
-    assert np.mean(fdp_list) <= alpha + test_tol
 
 
 def test_encluvi_temporal():
@@ -303,18 +149,21 @@ def test_encluvi_temporal():
             n_clusters=n_clusters, connectivity=connectivity, linkage="ward"
         )
 
-        cluvi = EnCluVI(
-            vi_estimator=DesparsifiedLasso(
-                estimator=MultiTaskLassoCV(max_iter=1000)
+        encludl = EnsembleImportance(
+            vim=ClusterImportance(
+                vim=DesparsifiedLasso(
+                    estimator=MultiTaskLassoCV(max_iter=1000)
+                ),
+                clustering=ward,
+                random_state=seed,
             ),
-            clustering=ward,
-            n_bootstraps=5,
+            n_repeats=5,
             random_state=seed,
         )
-        cluvi.fit_importance(X, y)
+        encludl.fit_importance(X, y)
 
         alpha = 0.1
-        selected = cluvi.fdr_selection(fdr=alpha, two_tailed_test=False)
+        selected = encludl.fdr_selection(fdr=alpha, two_tailed_test=False)
         gt_mask = np.zeros(n_features, dtype=int)
         gt_mask[:extended_support] = 1
         fdp, power = fdp_power(
@@ -341,11 +190,15 @@ def test_encluvi_independence():
     ward = FeatureAgglomeration(
         n_clusters=n_clusters, connectivity=connectivity, linkage="ward"
     )
-    encluvi = EnCluVI(
-        vi_estimator=DesparsifiedLasso(estimator=LassoCV()),
-        clustering=ward,
-        n_bootstraps=20,
-        cluster_bootstrap_size=0.5,
+
+    encluvi = EnsembleImportance(
+        vim=ClusterImportance(
+            vim=DesparsifiedLasso(estimator=LassoCV()),
+            clustering=ward,
+            random_state=1,
+        ),
+        bootstrap_frac=0.5,
+        n_repeats=20,
         random_state=1,
         n_jobs=1,
     )
